feature: move the func ReverseSlice
